Replace orgs.User proxy model with users.User custom model
