+++ temba/auth_tweaks/tests.py
@@ -1,4 +1,4 @@
-from django.contrib.auth.models import User
+from django.contrib.auth import get_user_model
 
 from temba.tests import TembaTest
 
@@ -6,4 +6,4 @@ from temba.tests import TembaTest
 class UserTest(TembaTest):
     def test_user_model(self):
         long_username = "bob12345678901234567890123456789012345678901234567890@msn.com"
-        User.objects.create(username=long_username, email=long_username)
+        get_user_model().objects.create(username=long_username, email=long_username)
