+++ temba/notifications/models.py
@@ -1,6 +1,7 @@
 import logging
 from abc import abstractmethod
 
+from django.conf import settings
 from django.db import models
 from django.db.models import Q
 from django.urls import reverse
@@ -171,7 +172,7 @@ class Notification(models.Model):
     # types like channel alerts, it will be the UUID of an object.
     scope = models.CharField(max_length=36)
 
-    user = models.ForeignKey(User, on_delete=models.PROTECT, related_name="notifications")
+    user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.PROTECT, related_name="notifications")
     is_seen = models.BooleanField(default=False)
     email_address = models.EmailField(null=True)  # only used when dest email != current user email
     email_status = models.CharField(choices=EMAIL_STATUS_CHOICES, max_length=1, default=EMAIL_STATUS_NONE)
