+++ temba/orgs/views/tests/test_contextprocessor.py
@@ -10,25 +10,25 @@ class OrgContextProcessorTest(TembaTest):
         self.assertTrue(perms["msgs"]["msg_list"])
         self.assertTrue(perms["contacts"]["contact_update"])
         self.assertTrue(perms["orgs"]["org_country"])
-        self.assertTrue(perms["orgs"]["user_list"])
         self.assertTrue(perms["orgs"]["org_delete"])
         self.assertTrue(perms["tickets"]["ticket_list"])
+        self.assertTrue(perms["users"]["user_list"])
 
         perms = RolePermsWrapper(OrgRole.EDITOR)
 
         self.assertTrue(perms["msgs"]["msg_list"])
         self.assertTrue(perms["contacts"]["contact_update"])
-        self.assertFalse(perms["orgs"]["user_list"])
         self.assertFalse(perms["orgs"]["org_delete"])
         self.assertTrue(perms["tickets"]["ticket_list"])
+        self.assertFalse(perms["users"]["user_list"])
 
         perms = RolePermsWrapper(OrgRole.AGENT)
 
         self.assertFalse(perms["msgs"]["msg_list"])
         self.assertFalse(perms["contacts"]["contact_update"])
-        self.assertFalse(perms["orgs"]["user_list"])
         self.assertFalse(perms["orgs"]["org_delete"])
         self.assertTrue(perms["tickets"]["ticket_list"])
+        self.assertFalse(perms["users"]["user_list"])
 
         self.assertFalse(perms["msgs"]["foo"])  # no blow up if perm doesn't exist
         self.assertFalse(perms["chickens"]["foo"])  # or app doesn't exist
